Move the facet number iterators into their own module
